// src/types.ts
/*
 * Copyright (c) 2019 - 2021 Dr. Krusche & Partner PartG. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not
 * use this file except in compliance with the License. You may obtain a copy of
 * the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations under
 * the License.
 *
 * @author Stefan Krusche, Dr. Krusche & Partner PartG
 *
 */
import { DataQuery, DataSourceJsonData } from '@grafana/data';

export interface WorksQuery extends DataQuery {
  dataset?: string;
  query?: string;
  type?: string;
}

/**
 * The enum values defined below refer to the GDELT visualization
 * project and must be customized to match other project demands.
 */
export enum WorksQueryType {
  getEvents = 'getEvents',
  getMentions = 'getMentions',
}

/**
 * These are options configured for each DataSource instance
 */
export interface WorksOptions extends DataSourceJsonData {
  path?: string;
}

/**
 * Value that is used in the backend, but never sent over HTTP to the frontend
 */
export interface WorksSecureJsonData {
  apiKey?: string;
}
/**
 * Datset specifies a network that is provdied by the PredictiveWorks.
 * backend. It is similar to a table in a relational database.
 */
export interface Dataset {
  index: string;
  label: string;
  value: string;
}

export interface Edge {
  id: string;
  src: string;
  dst: string;
  mainStat?: string;
}

export interface Node {
  id: string;
  title: string;
  subTitle?: string;
  mainStat?: number;
}